Errors with the code, it will not work for the card effects that last multiple turns. I am rewritting the script to allow for cards to have several effects, that take place over several turns. This is not super easy, and is not currently working, but just wanted to save it.

+++ Assets/Scripts/EconomyManager.cs
@@ -7,7 +7,7 @@ using TMPro;
 public class EconomyManager : MonoBehaviour
 {
     //Effectable Values
-    public int currentGold;
+    public float currentGold;
     public int currentSilver;
     public float currentTechPoints;
     public int currentWarWeariness;
